Updated Patient Homepage, connected PatientNearbyButton
